requirements.txt
langchain_googke_vertexai
langchain
python-dotenv
google-generativeai